better light

+++ src/PlayerMob.h
@@ -35,7 +35,7 @@ public:
   static PlayerMob* ply;
   static int distPlyCart(CPoint p); // right-angle distance.raw/true calc.
   static int distPly(CPoint p); // raw/true calc. (square)
-  static int distPlyLight(CPoint p); // light-adjusted
+  static double distPlyLight(CPoint p); // light-adjusted
 
   static PlayerMob* createPlayer();
 };
